requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
altair==6.0.0
altgraph==0.17.5
amply==0.1.6
annotated-types==0.6.0
anthropic==0.75.0
anyio==4.12.0
asttokens==2.4.1
astunparse==1.6.3
attr==0.3.2
attrs==23.2.0
audioread==3.0.1
awscli==1.33.44
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boltons==24.0.0
boto3==1.34.162
botocore==1.34.162
Brotli==1.1.0
brotlicffi==1.1.0.0
cached-property==1.5.2
cachetools==6.2.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configparser==7.0.0
contourpy==1.2.1
cryptography==42.0.5
curio==1.6
cycler==0.12.1
dataclasses-json==0.6.5
dateparser==1.1.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.16
einops==0.8.0
einops-exts==0.0.4
email_validator==2.1.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.51.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
funcsigs==1.0.2
future==1.0.0
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
google_search_results==2.4.2
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.63.0
gruut==2.3.4
gruut-ipa==0.13.0
gruut_lang_en==2.0.0
h11==0.16.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
html5lib==1.1
HTMLParser==0.0.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.0
hyperframe==6.0.1
hypothesis==6.100.2
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.3
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
keras==3.3.3
keyring==25.2.0
kiwisolver==1.4.5
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-community==0.0.38
langchain-core==1.2.5
langchain-text-splitters==0.0.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
lockfile==0.12.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==2.14.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
ordereddict==1.1
orjson==3.10.3
ormsgpack==1.12.1
otoole==1.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
pefile==2024.8.26
pickleshare==0.7.5
pillow==10.3.0
pipdeptree==2.19.1
pipreqs==0.5.0
platformdirs==4.2.1
pluggy==1.5.0
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==22.0.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydeps==1.12.20
pydot==4.0.0
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2025.11
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.3
PyQt6_sip==13.10.3
pytest==8.2.0
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.2
railroad==0.5.0
referencing==0.35.0
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
setuptools-scm==8.0.4
simple-websocket==1.0.0
six==1.16.0
slack_sdk==3.27.1
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
socksio==1.0.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
stack-data==0.6.3
stdlib-list==0.10.0
streamlit==1.52.2
sympy==1.12
tavily-python==0.7.17
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uuid_utils==0.12.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.6.0
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.43
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
zstandard==0.25.0
